pyproject.toml
[project]
name = "content-review"
version = "0.1.0"
description = "LangChain-based content moderation orchestrator (ReviewOrchestrator)"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "aiofiles>=23.0",
    "httpx>=0.27",
    # Pin before chromadb resolves deps: chromadb needs numpy; with cffi/cryptography for arch-matched wheels
    "cffi>=1.16.0",
    "cryptography>=42.0.0",
    "numpy>=1.26.0,<3",
    "chromadb>=0.5.0",
    "faiss-cpu>=1.8",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "click>=8.0",
    "python-multipart>=0.0.9",
    # Some deps (Chroma / vector stack) import jwt at runtime; declare explicitly to avoid missing installs
    "PyJWT>=2.8.0",
    "pypinyin>=0.50.0",
    "textual>=0.47.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "ImageHash>=4.3.1",
]

# Image pipeline (OCR via EasyOCR, pulls torch; first run downloads recognition models)
image = [
    "pillow>=10.0",
    "ImageHash>=4.3.1",
    "opencv-python-headless>=4.10",
    "easyocr>=1.7.0",
]

# Video pipeline
video = [
    "moviepy>=1.0",
    "openai-whisper>=20240930",
]

# Local llama.cpp models
local = [
    "llama-cpp-python>=0.2",
]

# Back-compat for `pip install .[tui]` (textual is already a core dep; this extra adds nothing)
tui = []

[project.scripts]
content-review = "reviewagent.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel>=0.38"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["reviewagent*"]

[tool.setuptools.package-data]
reviewagent = ["web/static/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"
